Add Period Model
I couldn't name the table "times", because rails reserves the keyword "Time" which is what our model would have been named. I think the name "Period" is suitable.

There definitely is room to modify the type and location to be foreign key references to another table of the respective column, but for starters, I think it's not completely necessary. When we start needing the added improvements, it's definitely a change we can incoporate.

For example, if we allow users to find classes based on location, it'd be in our best interest to make a separate location table with, at the very least, 2 columns: bldg and room_no. For now, it's unecessary complexity

Fields:
	- start_time (time)
	- end_time (time)
	- days (string)
	- type (string)
	- location (string)